+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/storage/keyvalue/KeyValueStorageProvider.java
@@ -69,11 +69,14 @@ public class KeyValueStorageProvider implements StorageProvider {
 
   @Override
   public BlockchainStorage createBlockchainStorage(
-      final ProtocolSchedule protocolSchedule, final VariablesStorage variablesStorage) {
+      final ProtocolSchedule protocolSchedule,
+      final VariablesStorage variablesStorage,
+      final DataStorageConfiguration dataStorageConfiguration) {
     return new KeyValueStoragePrefixedKeyBlockchainStorage(
         getStorageBySegmentIdentifier(KeyValueSegmentIdentifier.BLOCKCHAIN),
         variablesStorage,
-        ScheduleBasedBlockHeaderFunctions.create(protocolSchedule));
+        ScheduleBasedBlockHeaderFunctions.create(protocolSchedule),
+        dataStorageConfiguration.getReceiptCompactionEnabled());
   }
 
   @Override