+++ ethereum/eth/src/test/java/org/hyperledger/besu/ethereum/eth/sync/snapsync/StackTrieTest.java
@@ -18,12 +18,15 @@ import org.hyperledger.besu.datatypes.Hash;
 import org.hyperledger.besu.ethereum.core.TrieGenerator;
 import org.hyperledger.besu.ethereum.proof.WorldStateProofProvider;
 import org.hyperledger.besu.ethereum.trie.MerkleTrie;
+import org.hyperledger.besu.ethereum.trie.RangeManager;
 import org.hyperledger.besu.ethereum.trie.RangeStorageEntriesCollector;
 import org.hyperledger.besu.ethereum.trie.TrieIterator;
 import org.hyperledger.besu.ethereum.trie.forest.storage.ForestWorldStateKeyValueStorage;
+import org.hyperledger.besu.ethereum.trie.patricia.StoredMerklePatriciaTrie;
 import org.hyperledger.besu.ethereum.worldstate.WorldStateStorageCoordinator;
 import org.hyperledger.besu.services.kvstore.InMemoryKeyValueStorage;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.TreeMap;
 
@@ -141,6 +144,129 @@ public class StackTrieTest {
 
     Assertions.assertThat(
             worldStateKeyValueStorage.getAccountStateTrieNode(accountStateTrie.getRootHash()))
+        .isPresent();
+  }
+
+  @Test
+  public void shouldNotSaveNodeWithChildNotInTheRange() {
+    final ForestWorldStateKeyValueStorage worldStateStorage =
+        new ForestWorldStateKeyValueStorage(new InMemoryKeyValueStorage());
+    final WorldStateStorageCoordinator worldStateStorageCoordinator =
+        new WorldStateStorageCoordinator(worldStateStorage);
+
+    final MerkleTrie<Bytes, Bytes> trie =
+        new StoredMerklePatriciaTrie<>(
+            (location, hash) -> worldStateStorage.getAccountStateTrieNode(hash).map(Bytes::wrap),
+            b -> b,
+            b -> b);
+
+    trie.put(Bytes32.rightPad(Bytes.of(0x10)), Bytes.of(0x01));
+    trie.put(Bytes32.rightPad(Bytes.of(0x11)), Bytes.of(0x01));
+    trie.put(Bytes32.rightPad(Bytes.of(0x20)), Bytes.of(0x01));
+    trie.put(Bytes32.rightPad(Bytes.of(0x21)), Bytes.of(0x01));
+    trie.put(Bytes32.rightPad(Bytes.of(0x01)), Bytes.of(0x02));
+    trie.put(Bytes32.rightPad(Bytes.of(0x02)), Bytes.of(0x03));
+    trie.put(Bytes32.rightPad(Bytes.of(0x03)), Bytes.of(0x04));
+
+    final ForestWorldStateKeyValueStorage.Updater updater = worldStateStorage.updater();
+    trie.commit((location, hash, value) -> updater.putAccountStateTrieNode(hash, value));
+    updater.commit();
+
+    final Bytes32 startRange = Bytes32.rightPad(Bytes.of(0x02));
+
+    final RangeStorageEntriesCollector collector =
+        RangeStorageEntriesCollector.createCollector(
+            startRange, RangeManager.MAX_RANGE, 15, Integer.MAX_VALUE);
+    final TrieIterator<Bytes> visitor = RangeStorageEntriesCollector.createVisitor(collector);
+    final TreeMap<Bytes32, Bytes> entries =
+        (TreeMap<Bytes32, Bytes>)
+            trie.entriesFrom(
+                root ->
+                    RangeStorageEntriesCollector.collectEntries(
+                        collector, visitor, root, startRange));
+
+    final WorldStateProofProvider worldStateProofProvider =
+        new WorldStateProofProvider(worldStateStorageCoordinator);
+
+    // generate the proof
+    final List<Bytes> proofs =
+        worldStateProofProvider.getAccountProofRelatedNodes(
+            Hash.wrap(trie.getRootHash()), startRange);
+    proofs.addAll(
+        worldStateProofProvider.getAccountProofRelatedNodes(
+            Hash.wrap(trie.getRootHash()), entries.lastKey()));
+
+    // try to commit with stack trie
+    final ForestWorldStateKeyValueStorage recreatedWorldStateStorage =
+        new ForestWorldStateKeyValueStorage(new InMemoryKeyValueStorage());
+    final StackTrie stackTrie = new StackTrie(Hash.wrap(trie.getRootHash()), 0, 256, startRange);
+    stackTrie.addSegment();
+    stackTrie.addElement(Bytes32.random(), proofs, entries);
+    final ForestWorldStateKeyValueStorage.Updater updaterStackTrie =
+        recreatedWorldStateStorage.updater();
+    stackTrie.commit(
+        (location, hash, value) -> updaterStackTrie.putAccountStateTrieNode(hash, value));
+    updaterStackTrie.commit();
+
+    // verify the state of the db
+    Assertions.assertThat(worldStateStorage.getAccountStateTrieNode(trie.getRootHash()))
+        .isPresent();
+    Assertions.assertThat(recreatedWorldStateStorage.getAccountStateTrieNode(trie.getRootHash()))
+        .isNotPresent();
+  }
+
+  @Test
+  public void shouldSaveNodeWithAllChildsInTheRange() {
+    final ForestWorldStateKeyValueStorage worldStateStorage =
+        new ForestWorldStateKeyValueStorage(new InMemoryKeyValueStorage());
+
+    final MerkleTrie<Bytes, Bytes> trie =
+        new StoredMerklePatriciaTrie<>(
+            (location, hash) -> worldStateStorage.getAccountStateTrieNode(hash).map(Bytes::wrap),
+            b -> b,
+            b -> b);
+
+    trie.put(Bytes32.rightPad(Bytes.of(0x10)), Bytes.of(0x01));
+    trie.put(Bytes32.rightPad(Bytes.of(0x11)), Bytes.of(0x01));
+    trie.put(Bytes32.rightPad(Bytes.of(0x20)), Bytes.of(0x01));
+    trie.put(Bytes32.rightPad(Bytes.of(0x21)), Bytes.of(0x01));
+    trie.put(Bytes32.rightPad(Bytes.of(0x01)), Bytes.of(0x02));
+    trie.put(Bytes32.rightPad(Bytes.of(0x02)), Bytes.of(0x03));
+    trie.put(Bytes32.rightPad(Bytes.of(0x03)), Bytes.of(0x04));
+
+    final ForestWorldStateKeyValueStorage.Updater updater = worldStateStorage.updater();
+    trie.commit((location, hash, value) -> updater.putAccountStateTrieNode(hash, value));
+    updater.commit();
+
+    final Bytes32 startRange = Bytes32.rightPad(Bytes.of(0x00));
+
+    final RangeStorageEntriesCollector collector =
+        RangeStorageEntriesCollector.createCollector(
+            startRange, RangeManager.MAX_RANGE, 15, Integer.MAX_VALUE);
+    final TrieIterator<Bytes> visitor = RangeStorageEntriesCollector.createVisitor(collector);
+    final TreeMap<Bytes32, Bytes> entries =
+        (TreeMap<Bytes32, Bytes>)
+            trie.entriesFrom(
+                root ->
+                    RangeStorageEntriesCollector.collectEntries(
+                        collector, visitor, root, startRange));
+
+    // try to commit with stack trie
+    final ForestWorldStateKeyValueStorage recreatedWorldStateStorage =
+        new ForestWorldStateKeyValueStorage(new InMemoryKeyValueStorage());
+    final StackTrie stackTrie = new StackTrie(Hash.wrap(trie.getRootHash()), 0, 256, startRange);
+    stackTrie.addSegment();
+    stackTrie.addElement(Bytes32.random(), new ArrayList<>(), entries);
+    final ForestWorldStateKeyValueStorage.Updater updaterStackTrie =
+        recreatedWorldStateStorage.updater();
+    stackTrie.commit(
+        (location, hash, value) -> updaterStackTrie.putAccountStateTrieNode(hash, value));
+    updaterStackTrie.commit();
+
+    // verify the state of the db
+    Assertions.assertThat(worldStateStorage.getAccountStateTrieNode(trie.getRootHash()))
+        .isPresent();
+    Assertions.assertThat(recreatedWorldStateStorage.getAccountStateTrieNode(trie.getRootHash()))
         .isPresent();
   }
 }