+++ plugins/rocksdb/src/main/java/org/hyperledger/besu/plugin/services/storage/rocksdb/RocksDBKeyValuePrivacyStorageFactory.java
@@ -46,7 +46,9 @@ public class RocksDBKeyValuePrivacyStorageFactory implements PrivacyKeyValueStor
   private static final Set<PrivacyVersionedStorageFormat> SUPPORTED_VERSIONS =
       EnumSet.of(
           PrivacyVersionedStorageFormat.FOREST_WITH_VARIABLES,
-          PrivacyVersionedStorageFormat.BONSAI_WITH_VARIABLES);
+          PrivacyVersionedStorageFormat.FOREST_WITH_RECEIPT_COMPACTION,
+          PrivacyVersionedStorageFormat.BONSAI_WITH_VARIABLES,
+          PrivacyVersionedStorageFormat.BONSAI_WITH_RECEIPT_COMPACTION);
   private static final String PRIVATE_DATABASE_PATH = "private";
   private final RocksDBKeyValueStorageFactory publicFactory;
   private DatabaseMetadata databaseMetadata;
@@ -145,7 +147,8 @@ public class RocksDBKeyValuePrivacyStorageFactory implements PrivacyKeyValueStor
         privacyMetadata = existingPrivacyMetadata;
         final int existingPrivacyVersion = maybeExistingPrivacyVersion.getAsInt();
         final var runtimeVersion =
-            PrivacyVersionedStorageFormat.defaultForNewDB(commonConfiguration.getDatabaseFormat());
+            PrivacyVersionedStorageFormat.defaultForNewDB(
+                commonConfiguration.getDataStorageConfiguration());
 
         if (existingPrivacyVersion > runtimeVersion.getPrivacyVersion().getAsInt()) {
           final var maybeDowngradedMetadata =
@@ -194,6 +197,16 @@ public class RocksDBKeyValuePrivacyStorageFactory implements PrivacyKeyValueStor
     // In case we do an automated downgrade, then we also need to update the metadata on disk to
     // reflect the change to the runtime version, and return it.
 
+    // Besu supports both formats of receipts so no downgrade is needed
+    if (runtimeVersion == PrivacyVersionedStorageFormat.BONSAI_WITH_VARIABLES
+        || runtimeVersion == PrivacyVersionedStorageFormat.FOREST_WITH_VARIABLES) {
+      LOG.warn(
+          "Database contains compacted receipts but receipt compaction is not enabled, new receipts  will "
+              + "be not stored in the compacted format. If you want to remove compacted receipts from the "
+              + "database it is necessary to resync Besu. Besu can support both compacted and non-compacted receipts.");
+      return Optional.empty();
+    }
+
     // for the moment there are supported automated downgrades, so we just fail.
     String error =
         String.format(
@@ -215,6 +228,18 @@ public class RocksDBKeyValuePrivacyStorageFactory implements PrivacyKeyValueStor
     // database.
     // In case we do an automated upgrade, then we also need to update the metadata on disk to
     // reflect the change to the runtime version, and return it.
+
+    // Besu supports both formats of receipts so no upgrade is needed other than updating metadata
+    if (runtimeVersion == PrivacyVersionedStorageFormat.BONSAI_WITH_RECEIPT_COMPACTION
+        || runtimeVersion == PrivacyVersionedStorageFormat.FOREST_WITH_RECEIPT_COMPACTION) {
+      final DatabaseMetadata metadata = new DatabaseMetadata(runtimeVersion);
+      try {
+        metadata.writeToDirectory(dataDir);
+        return Optional.of(metadata);
+      } catch (IOException e) {
+        throw new StorageException("Database upgrade to use receipt compaction failed", e);
+      }
+    }
 
     // for the moment there are no planned automated upgrades, so we just fail.
     String error =